add feature of email lifespan and some css styles

+++ WebSite/main_page/views.py
@@ -1,6 +1,5 @@
 from django.views import View
 from django.shortcuts import render, redirect
-from django.utils import timezone
 from core_functionality.main import setup, delete_mail
 from .models import UserEmail
 
@@ -9,27 +8,28 @@ class EmailView(View):
     template_name = 'main_page/email_generator.html'
 
     def get(self, request):
+        # Получаем текущий email, если он есть и не истек
         email_record = UserEmail.objects.filter().first()
-        if email_record:
-            if email_record.is_expired():
-                delete_mail(mail=email_record.email)
-                email_record.delete()
-                email_record = None
+        if email_record and email_record.is_expired():
+            delete_mail(mail=email_record.email)
+            email_record.delete()
+            email_record = None
 
-        if not email_record:
-            email = setup()
-            email_record = UserEmail.objects.create(email=email)
-
-        return render(request, self.template_name, {'email': email_record.email})
+        # Передаем текущий email, если он есть
+        return render(request, self.template_name, {'email': email_record.email if email_record else None})
 
     def post(self, request):
+        lifespan = int(request.POST.get('lifespan', 2))  # Получаем выбранное время жизни
         email_record = UserEmail.objects.filter().first()
+
+        # Удаляем старый email, если есть
         if email_record:
             delete_mail(mail=email_record.email)
             email_record.delete()
 
+        # Генерируем новый email и сохраняем его с выбранным временем жизни
         email = setup()
-        UserEmail.objects.create(email=email)
+        UserEmail.objects.create(email=email, lifespan_minutes=lifespan)
         return redirect('/')
